MFC:   Understand environment variables in commands   Understand IP6 addresses in ``show route''   Correct an alignment bug in ``show route''   Add the ``resolv'' command   Add ``dns'' logging   Rename struct descriptor -> struct fdescriptor   Fix some printf-style parameter-type bugs   Use <nnn> instead of <???> when something's not understood   Log (debug) information about incoming fragment handling   Correct two bugs in the timer code   Add some {Net,Open}BSD compatibility stuff   Fix some gramar and line breaks in ppp.8
Requested by: peter

+++ usr.sbin/ppp/mbuf.c
@@ -112,8 +112,8 @@ m_get(size_t m_len, int type)
      */
     *mb = (struct mbucket *)malloc(BUCKET_CHUNK * size);
     if (*mb == NULL) {
-      log_Printf(LogALERT, "Failed to allocate memory (%u)\n",
-                 BUCKET_CHUNK * size);
+      log_Printf(LogALERT, "Failed to allocate memory (%lu)\n",
+                 (unsigned long)BUCKET_CHUNK * size);
       AbortProgram(EX_OSERR);
     }
     bp = &(*mb)->u.m;
@@ -354,7 +354,7 @@ m_enqueue(struct mqueue *queue, struct mbuf *bp)
     } else
       queue->last = queue->top = bp;
     queue->len++;
-    log_Printf(LogDEBUG, "m_enqueue: len = %d\n", queue->len);
+    log_Printf(LogDEBUG, "m_enqueue: len = %lu\n", (unsigned long)queue->len);
   }
 }
